[refactor] #50 - 컴포넌트 리팩토링
